@@ -7,14 +7,14 @@ from click.testing import CliRunner
 
 from todowrite.app import ToDoWrite
 from todowrite.cli import cli
-from todowrite.db.models import (Artifact, Command, Label, Link, Node,
-                                 node_labels)
+from todowrite.db.models import Artifact, Command, Label, Link, Node, node_labels
 
 
 class TestCli(unittest.TestCase):
+    app: ToDoWrite
 
     @classmethod
-    def setUpClass(cls):
+    def setUpClass(cls) -> None:
         """Start the PostgreSQL container."""
         subprocess.run(["docker-compose", "up", "-d"], check=True)
         # Wait for the database to be ready
@@ -23,11 +23,11 @@ class TestCli(unittest.TestCase):
         cls.app = ToDoWrite(db_url)
 
     @classmethod
-    def tearDownClass(cls):
+    def tearDownClass(cls) -> None:
         """Stop the PostgreSQL container."""
         subprocess.run(["docker-compose", "down"], check=True)
 
-    def setUp(self):
+    def setUp(self) -> None:
         self.runner = CliRunner()
         # The CLI will use the environment variable for the database URL
         os.environ["TODOWRITE_DATABASE_URL"] = (
@@ -35,7 +35,7 @@ class TestCli(unittest.TestCase):
         )
         self.app.init_database()
 
-    def tearDown(self):
+    def tearDown(self) -> None:
         session = self.app.Session()
         session.execute(node_labels.delete())
         session.query(Artifact).delete()
@@ -46,13 +46,13 @@ class TestCli(unittest.TestCase):
         session.commit()
         session.close()
 
-    def test_init_command(self):
+    def test_init_command(self) -> None:
         """Test the init command."""
         result = self.runner.invoke(cli, ["init"])
         self.assertEqual(result.exit_code, 0)
         self.assertEqual(result.output, "Database initialized.\n")
 
-    def test_create_command(self):
+    def test_create_command(self) -> None:
         """Test the create command."""
         result = self.runner.invoke(cli, ["init"])
         self.assertEqual(result.exit_code, 0)
@@ -66,7 +66,7 @@ class TestCli(unittest.TestCase):
         self.assertEqual(result.exit_code, 0)
         self.assertIn("Node created:", result.output)
 
-    def test_get_command(self):
+    def test_get_command(self) -> None:
         """Test the get command."""
         result = self.runner.invoke(cli, ["init"])
         self.assertEqual(result.exit_code, 0)
@@ -84,7 +84,7 @@ class TestCli(unittest.TestCase):
         self.assertEqual(result.exit_code, 0)
         self.assertIn(f"ID: {node_id}", result.output)
 
-    def test_list_command(self):
+    def test_list_command(self) -> None:
         """Test the list command."""
         result = self.runner.invoke(cli, ["init"])
         self.assertEqual(result.exit_code, 0)
